tweaks + fixes

+++ sparsebak.py
@@ -5,25 +5,24 @@
 
 
 
-import re
-import os
-import time
-import mmap
-import subprocess
+import os, shutil, subprocess, time
+from pathlib import Path
+import re, mmap, lzma
 import xml.etree.ElementTree
 from optparse import OptionParser
 #import qubesadmin.tools
 
-bkset = "set01"
+bkset = "set04"
 setdir = "/baktest/"+bkset
 vgname = "qubes_dom0"
 poolname = "pool00"
 
-volfile = "/tmp/sparsebak-volumes.txt"
-deltafile = "/tmp/sparsebak-delta."
-dtree = None
-dblocksize = None
+tmpdir = "/tmp/sparsebak"
+volfile = tmpdir+"/volumes.txt"
+deltafile = tmpdir+"/delta."
 allvols = {}
+datavols = []
+newvols = []
 bkdir = setdir+"/"+vgname+"%"+poolname
 bksession = time.strftime("S_%Y%m%d-%H%M%S")
 bs = 512
@@ -45,6 +44,20 @@ parser.add_option("-u", "--unattended", action="store_true", dest="unattended", 
 monitor_only = not options.send # gather metadata without backing up if True
 
 
+
+def get_configs():
+    with open(setdir+"/sparsebak.conf", "r") as f:
+        datavols = f.read().splitlines()
+    return datavols
+
+
+# Determine previous session, and if it completed.
+# Example: if .deltamap-tmp exists, then perform checks on
+# which snapshots exist.
+def detect_state():
+    pass
+
+
 ## TICK - process metadata and compile delta info
 def prepare_snapshots():
 
@@ -59,39 +72,48 @@ def prepare_snapshots():
     for datavol in datavols:
         snap1vol = datavol + ".tick"
         snap2vol = datavol + ".tock"
-        try:
-            p = subprocess.check_output( ["lvs", vgname+"/"+datavol],
-                                        stderr=subprocess.STDOUT )
-        except:
+        if not lv_exists(vgname, datavol):
             # FIX: Remove non-existent volume from list...........
             print("ERROR:", datavol, "does not exist!")
             exit(1) #fix
 
-        # Make initial snapshot if necessary: ##FIX - this also means creating a new set!
-        try:
-            p = subprocess.check_output( ["lvs", vgname+"/"+snap1vol],
-                                        stderr=subprocess.STDOUT)
-        except:
-            p = subprocess.check_output(["lvcreate", "-pr", "-kn", "-ay", "-s",
-                vgname+"/"+datavol, "-n", snap1vol], stderr=subprocess.STDOUT)
-            print("Initial snapshot created:", snap1vol)
-
         # Check for stale snap2vol
-        try:
-            p = subprocess.check_output( ["lvs", vgname+"/"+snap2vol],
-                                        stderr=subprocess.STDOUT)
-        except:
-            pass
-        else:
+        if lv_exists(vgname, snap2vol):
             p = subprocess.check_output(["lvremove", "-f",vgname+"/"+snap2vol],
                                         stderr=subprocess.STDOUT)
             print("Stale snapshot removed.")
             # also cleanup other data from aborted backup...
 
+        # Make initial snapshot if necessary:
+        if not os.path.exists(bkdir+"/"+datavol+".deltamap") \
+            and not os.path.exists(bkdir+"/"+datavol+".deltamap-tmp"):
+            if options.full:
+                p = subprocess.check_output(["lvcreate", "-pr", "-kn", "-ay", "-s",
+                    vgname+"/"+datavol, "-n", snap1vol], stderr=subprocess.STDOUT)
+                print("Initial snapshot created:", snap1vol)
+            newvols.append(datavol)
+        elif os.path.exists(bkdir+"/"+datavol+".deltamap-tmp"):
+            raise Exception("ERROR - Previous delta map was not finalized. Rename "
+                +mapstate+"-tmp without -tmp to recover.")
+            # Fix: ask to recover/use the tmp file
+        elif not lv_exists(vgname, snap1vol):
+            raise Exception("ERROR: Map and snapshots in inconsistent state, "
+                +snap1vol+" is missing!")
+
         # Make current snapshot
         p = subprocess.check_output( ["lvcreate", "-pr", "-kn", "-ay", \
             "-s", vgname+"/"+datavol, "-n",snap2vol], stderr=subprocess.STDOUT)
-        print("Tracking snapshot created:", snap2vol)
+        print("Current snapshot created:", snap2vol)
+
+
+def lv_exists(vgname, lvname):
+    try:
+        p = subprocess.check_output( ["lvs", vgname+"/"+lvname],
+                                    stderr=subprocess.STDOUT )
+    except:
+        return False
+    else:
+        return True
 
 
 # Load lvm metadata
@@ -102,7 +124,7 @@ def get_lvm_metadata():
     with open(volfile) as f:
         lines = f.readlines()
     scope = 0
-    volume = ""
+    volume = devid = trans = ""
     version = False
     for l in lines:
         if l.strip() == "version = 1":
@@ -118,28 +140,54 @@ def get_lvm_metadata():
         scope -= l.count('}')
         if scope == 3 and not refind == l:
             volume = refind.strip()
-        elif scope == 4 and not volume == "" and "device_id =" in l:
-## Also pull transaction_id and compare datavol vs snap1 freshness b4 analyzing ##
-            allvols[volume] = re.sub("device_id = ([0-9]+)", r'\1', l).strip()
-            volume = ""
+            allvols[volume] = [None, None]
+        elif scope == 4 and volume > "" and None in allvols[volume] and "device_id =" in l:
+            devid = re.sub("device_id = ([0-9]+)", r'\1', l).strip()
+        elif scope == 4 and volume > "" and None in allvols[volume] and "transaction_id =" in l:
+            trans = re.sub("transaction_id = ([0-9]+)", r'\1', l).strip()
         elif scope == 0 and '}' in l:
             break
+        if devid > "" and trans > "":
+            allvols[volume] = [devid, trans]
+            volume = devid = trans = ""
 
-def get_lvm_sizes():
+
+def get_lvm_size(vol):
     line = subprocess.check_output( ["lvdisplay --units=b " \
-        + " /dev/mapper/"+vgname+"-"+snap1vol.replace("-","--") \
+        + " /dev/mapper/"+vgname+"-"+vol.replace("-","--") \
         +  "| grep 'LV Size'"], shell=True).decode("UTF-8").strip()
-    snap1size = int(re.sub("^.+ ([0-9]+) B", r'\1', line))
-    line = subprocess.check_output( ["lvdisplay --units=b " \
-        + " /dev/mapper/"+vgname+"-"+snap2vol.replace("-","--") \
-        +  "| grep 'LV Size'"], shell=True).decode("UTF-8").strip()
-    snap2size = int(re.sub("^.+ ([0-9]+) B", r'\1', line))
-    return snap1size, snap2size
+    return int(re.sub("^.+ ([0-9]+) B", r'\1', line))
+
+
+# Compare transaction ids, include only vols with new or doing first backup
+def find_changed_vols():
+    dvs = []
+    for datavol in datavols:
+        snap1vol = datavol + ".tick"
+        snap2vol = datavol + ".tock"
+        reason = ""
+        if datavol in newvols and options.full:
+            continue
+        elif datavol in newvols:
+            reason = "- needs initial full backup"
+        elif int(allvols[datavol][1]) <= int(allvols[snap1vol][1]):
+            reason = "- no new transactions"
+        else:
+            dvs.append(datavol)
+
+        if reason > "":
+            print("  Skipping", datavol, reason)
+            p = subprocess.check_output(["lvremove", "-f",vgname+"/"+snap2vol],
+                                        stderr=subprocess.STDOUT)
+
+    return dvs
 
 
 # Get delta between snapshots
 def get_lvm_deltas():
-    print("Acquiring LVM delta info...")
+    print("\nAcquiring LVM delta info...")
+    subprocess.run(["dmsetup","message", vgname+"-"+poolname+"-tpool", \
+        "0", "release_metadata_snap"], check=False, stderr=subprocess.DEVNULL)
     subprocess.check_call( ["dmsetup", "message", vgname+"-"+poolname+"-tpool", \
         "0", "reserve_metadata_snap"] )
     td_err = False
@@ -149,8 +197,8 @@ def get_lvm_deltas():
         try:
             with open(deltafile+datavol, "w") as f:
                 cmd = ["thin_delta -m" \
-                    + " --thin1 " + allvols[snap1vol] \
-                    + " --thin2 " + allvols[snap2vol] \
+                    + " --thin1 " + allvols[snap1vol][0] \
+                    + " --thin2 " + allvols[snap2vol][0] \
                     + " /dev/mapper/"+vgname+"-"+poolname+"_tmeta" \
                     + " | grep -v '<same .*\/>$'" ]
                 #print(cmd[0])
@@ -165,6 +213,11 @@ def get_lvm_deltas():
 
 
 def update_delta_digest():
+
+    if not os.path.exists(mapstate):
+        print("Skipping map update.")
+        return False, False
+
     print("Updating block change map...")
     dtree = xml.etree.ElementTree.parse(deltafile+datavol).getroot()
     dblocksize = int(dtree.get("data_block_size"))
@@ -173,23 +226,12 @@ def update_delta_digest():
     dnewblocks = 0
     dfreedblocks = 0
 
-    dest = bkdir+"/"+datavol+"_deltamap.dat"
-    #bmap_size = int(snap2size / bkchunksize / 8) + 1
-    if not os.access(dest, os.F_OK):
-        print("Volume", datavol, "not initialized; Skipping map update.")
-        return False
-    os.rename(dest, dest+"-tmp")
-
-    with open(dest+"-tmp", "r+b") as bmapf:
+    os.rename(mapstate, mapstate+"-tmp")
+    with open(mapstate+"-tmp", "r+b") as bmapf:
         os.ftruncate(bmapf.fileno(), bmap_size)
         bmap_mm = mmap.mmap(bmapf.fileno(), 0)
 
         for delta in dtree.find("diff"):
-            #for i in dtree.find("diff").findall("right_only"):
-            #if delta.tag == "same":
-            #    continue
-            #elif delta.tag in ["different", "left_only", "right_only"]:
-
             blockbegin = int(delta.get("begin")) * dblocksize
             blocklen   = int(delta.get("length")) * dblocksize
             if delta.tag in ["different", "right_only"]:
@@ -212,17 +254,10 @@ def update_delta_digest():
                 seg = volsegment
 #        print("WRITE:")
         bmap_mm[lastindex] |= bmap_byte
-    os.rename(dest+"-tmp", dest)
-    print("Data changes:", dnewblocks * bs, "bytes")
-    print("Deallocated :", dfreedblocks * bs, "bytes")
-    return True
-
-
-def rotate_snapshots():
-    print("Rotating snapshots for", datavol)
-    ##FIX - Make this atomic:
-    p = subprocess.check_output( ["lvremove", "--force", vgname+"/"+snap1vol ])
-    p = subprocess.check_output( ["lvrename", vgname+"/"+snap2vol, snap1vol ])
+    #os.rename(mapstate+"-tmp", mapstate)
+    print("Data changed :", dnewblocks * bs, "bytes")
+    print("   Discarded :", dfreedblocks * bs, "bytes")
+    return True, dnewblocks+dfreedblocks > 0
 
 
 ## TOCK - Run backup session
@@ -230,19 +265,15 @@ def rotate_snapshots():
 def record_to_bkdir(send_all = False):
     # Refactor for realistic 'send' use cases
     # Add a stat file for each vol/session showing vol size and previous session
-    bmfile = bkdir+"/"+datavol+"_deltamap.dat"
-    if not send_all and not mapped:
-        print("Volume", datavol, "not initialized; Skipping backup.")
-        return False
-        
+
     destdir=bkdir+"/"+datavol+"/"+bksession
     print("Backing up to", destdir)
     os.makedirs(destdir+"-tmp")
     zeros = bytes(bkchunksize)
     bcount = 0
 
-    with open("/dev/mapper/"+vgname+"-"+snap2vol.replace("-","--"), "rb") as vf:
-        with open("/dev/zero" if send_all else bmfile, "r+b") as bmapf:
+    with open("/dev/mapper/"+vgname+"-"+snap2vol.replace("-","--"),"rb") as vf:
+        with open("/dev/zero" if send_all else mapstate+"-tmp","r+b") as bmapf:
             bmap_mm = bytes(1) if send_all else mmap.mmap(bmapf.fileno(), 0)
             for addr in range(0, snap2size, bkchunksize):
                 bmap_pos = int(addr / bkchunksize / 8)
@@ -252,72 +283,97 @@ def record_to_bkdir(send_all = False):
                     ## REVIEW int math vs large vol sizes . . .
                     vf.seek(addr)
                     buf = vf.read(bkchunksize)
-                    with open(destdir+"-tmp/" \
-                         + format(addr,"016x"), "wb") as segf:
-                        if buf != zeros: # write only non-empty
+                    destfile = destdir+"-tmp/"+format(addr,"016x")
+                    #subdir = destfile[:-7] ???
+                    if buf != zeros: # write only non-empty
+                        # Performance fix: move compression into separate processes
+                        with lzma.open(destfile, "wb", preset=0) as segf:
                             segf.write(buf)
-                            bcount += len(buf)
-                            print("DATA", end="")
+                        bcount += len(buf)
+                        print("DATA", end="")
+                    else:
+                        Path(destfile).touch()
                 print("     ", end="\x0d")
     os.rename(destdir+"-tmp", destdir)
     print("Bytes sent:", bcount)
     return True
 
 
-def finalize_monitor_session():
-    rotate_snapshots()
-
-
-def finalize_bk_session():
-    init_deltamap(bkdir+"/"+datavol+"_deltamap.dat")
-    rotate_snapshots()
-
-
-def get_configs():
-    with open(setdir+"/sparsebak.conf", "r") as f:
-        datavols = f.read().splitlines()
-    return datavols
-
-
 def init_deltamap(bmfile):
-    try:
+    if os.path.exists(bmfile):
         os.remove(bmfile)
-    except:
-        pass
+    if os.path.exists(bmfile+"-tmp"):
+        os.remove(bmfile+"-tmp")
     with open(bmfile, "wb") as bmapf:
         os.ftruncate(bmapf.fileno(), bmap_size)
 
 
+def rotate_snapshots():
+    print("Rotating snapshots for", datavol)
+    # Review: this should be atomic
+    p = subprocess.check_output( ["lvremove", "--force", vgname+"/"+snap1vol ])
+    p = subprocess.check_output( ["lvrename", vgname+"/"+snap2vol, snap1vol ])
 
-## Main ###############################################
+
+def finalize_monitor_session():
+    if map_updated:
+        rotate_snapshots()
+    os.rename(mapstate+"-tmp", mapstate)
+
+
+def finalize_bk_session():
+    rotate_snapshots()
+    init_deltamap(mapstate)
+
+
+
+## Main ###########################################################
 # ToDo: Check root user, thinp commands, etc.
 # Detect/make process lock
 # Check dir/file presence, volume sizes, deltabmap size
-# Check for modulus bkchunksize/(bs * dblocksize) == 0
+# Check for modulus bkchunksize/(bs * dblocksize) == 0 and mod 4096
+#
 # Make separate /var (local) and dest (remote) directories
 
 print("\nStarting", ["backup","monitor-only"][int(monitor_only)],"session:\n")
-if not os.access(bkdir, os.F_OK):
+
+shutil.rmtree(tmpdir+"-old", ignore_errors=True)
+if os.path.exists(tmpdir):
+    os.rename(tmpdir, tmpdir+"-old")
+os.makedirs(tmpdir)
+if not os.path.exists(bkdir):
     os.makedirs(bkdir)
 
 datavols = get_configs()
+detect_state()
 prepare_snapshots()
 get_lvm_metadata()
-get_lvm_deltas()
+datavols = find_changed_vols()
 
-for datavol in datavols:
+if not options.full:
+    newvols = []
+if len(datavols) > 0:
+    get_lvm_deltas()
+elif len(newvols) == 0:
+    print("No new data.")
+    exit(0)
+
+for datavol in datavols+newvols:
     print("\n** Starting", datavol)
     snap1vol = datavol + ".tick"
     snap2vol = datavol + ".tock"
-    snap1size, snap2size = get_lvm_sizes()
+    snap1size = get_lvm_size(snap1vol)
+    snap2size = get_lvm_size(snap2vol)
     bmap_size = int(snap2size / bkchunksize / 8) + 1
 
-    mapped = update_delta_digest()
+    mapstate = bkdir+"/"+datavol+".deltamap"
+    map_exists, map_updated = update_delta_digest()
 
     if not monitor_only:
-        if record_to_bkdir(send_all = not mapped and options.full):
+        if record_to_bkdir(send_all = datavol in newvols):
             finalize_bk_session()
     else:
         finalize_monitor_session()
+
 
 print("\nDone.\n")
